Moving taskPrototype to its own module - causes errors
